Use console_scripts
