+++ dockers/docker-fpm-frr/base_image_files/monit_bgp
@@ -20,7 +20,7 @@ check program bgp|bgpd with path "/usr/bin/process_checker bgp /usr/lib/frr/bgpd
 check program bgp|staticd with path "/usr/bin/process_checker bgp /usr/lib/frr/staticd"
     if status != 0 for 5 times within 5 cycles then alert
 
-check program bgp|bgpcfgd with path "/usr/bin/process_checker bgp /usr/bin/python -m bgpcfgd"
+check program bgp|bgpcfgd with path "/usr/bin/process_checker bgp /usr/bin/python /usr/local/bin/bgpcfgd"
     if status != 0 for 5 times within 5 cycles then alert
 
 check program bgp|bgpmon with path "/usr/bin/process_checker bgp /usr/bin/python /usr/local/bin/bgpmon"
