+++ dockers/docker-fpm-frr/frr/supervisord/supervisord.conf.j2
@@ -74,7 +74,7 @@ dependent_startup=true
 dependent_startup_wait_for=bgpd:running
 
 [program:bgpcfgd]
-command=/usr/bin/env python -m bgpcfgd
+command=/usr/bin/env python /usr/local/bin/bgpcfgd
 priority=6
 autostart=false
 autorestart=false
