+++ src/sonic-bgpcfgd/bgpcfgd/main.py
@@ -1,18 +1,26 @@
+import os
+import signal
+import sys
+import syslog
+import traceback
+
 from swsscommon import swsscommon
 
 from .config import ConfigMgr
 from .directory import Directory
+from .log import log_notice, log_crit
 from .managers_allow_list import BGPAllowListMgr
 from .managers_bgp import BGPPeerMgrBase
 from .managers_db import BGPDataBaseMgr
 from .managers_intf import InterfaceMgr
 from .managers_setsrc import ZebraSetSrc
-from .runner import Runner
+from .runner import Runner, signal_handler
 from .template import TemplateFabric
 from .utils import wait_for_daemons, read_constants
+from .vars import g_debug
 
 
-def main():
+def do_work():
     """ Main function """
     wait_for_daemons(["bgpd", "zebra", "staticd"], seconds=20)
     #
@@ -44,3 +52,30 @@ def main():
     for mgr in managers:
         runner.add_manager(mgr)
     runner.run()
+
+
+def main():
+    rc = 0
+    try:
+        syslog.openlog('bgpcfgd')
+        signal.signal(signal.SIGTERM, signal_handler)
+        signal.signal(signal.SIGINT, signal_handler)
+        do_work()
+    except KeyboardInterrupt:
+        log_notice("Keyboard interrupt")
+    except RuntimeError as exc:
+        log_crit(str(exc))
+        rc = -2
+        if g_debug:
+            raise
+    except Exception as exc:
+        log_crit("Got an exception %s: Traceback: %s" % (str(exc), traceback.format_exc()))
+        rc = -1
+        if g_debug:
+            raise
+    finally:
+        syslog.closelog()
+    try:
+        sys.exit(rc)
+    except SystemExit:
+        os._exit(rc)
