+++ src/sonic-bgpcfgd/bgpcfgd/vars.py
@@ -1,2 +1,1 @@
 g_debug = True  # FIXME: read from env variable, or from constants
-g_run = True
